+++ packages/spacecat-shared-data-access/src/models/organization.js
@@ -86,7 +86,7 @@ const Organization = (data = {}) => {
       throw new Error('Config must be provided');
     }
 
-    self.state.config = Config.toDynamoItem(config);
+    self.state.config = Config(config);
     self.touch();
 
     return self;